rework logic
instead of having each namespace hold a client list, just make it global so we only have to walk one list.
fix XnamespaceAssignByClientName to just use the name given.
move the client name check to ClientStateInitial because that much *can* be done without any fuss
the start of ClientStateRunning now checks for that, making the checks overall smoother.
actually assign the authId for the priv of new namespaces (whoops)
finally, truly, if all checks fail by the end of the ClientStateRunning hook,
the client can be considered "foreign". subj->authID will be 0 and no namespace will be assigned.